Refactor to increase sharing of gentle_overwrite and ensure that its behavior is shared between all producers of clvm code to disk
